% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{code_template}
\alias{code_template}
\title{Code template that is used by semproducible to generate R code}
\usage{
code_template()
}
\value{
a character string with the code template.
}
\description{
Use this template as a start for creating a custom template for your needs.
}
\details{
There are several variables in the template that is replaced with the
actual values. For instance:

%%FORMULA%% will be replaced with the lavaan formula.

%%OBSERVATIONS%% will be replaced with the number of observations of the
actual data.

%%TARGET%% will be replaced with the variable name of the covariance matrix.

%%VARIABLES%% will be replaced with the column names (variable names) of the
covariance matrix.

%%VALUES%% will be replaced with the actual values of the covariance matrix.
}
\examples{
\dontrun{
# Create random data.
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Modify code template.
template <- paste(code_template(), "# A comment placed at end of code")

# Generate R code with the template.
code <- semproducible(data, formula="y ~ x", code_template = template)
}
}
