% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.semproducible}
\alias{print.semproducible}
\title{Print semproducible code}
\usage{
\method{print}{semproducible}(x, ...)
}
\arguments{
\item{x}{semproducible code to print.}

\item{...}{other arguments passed on to the \code{\link[base:cat]{cat}}
function.}
}
\value{

}
\description{
Prints semproducible code to the console with colors.
}
\examples{
library(semproducible)

code <- semproducible(iris, drop_non_numeric=TRUE)

code
}
