% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{save_code}
\alias{save_code}
\title{Save code to a file}
\usage{
save_code(code, filename, overwrite = FALSE)
}
\arguments{
\item{code}{character string with R code.}

\item{filename}{character string with filename of the file to be saved.}
}
\description{
Save generated code to a file. Does not overwrite existing files unless
\code{overwrite} is set to \code{TRUE}.
}
\examples{
\dontrun{
# Create random data.
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Generate R code.
code <- semproducible(data, formula="y ~ x")

# Save R code to file.
save_code(code, "create_data.r")

# If the file already exists, you will get an error.
# You need to explicitly overwrite existing files.
save_code(code, "create_data.r", overwrite = TRUE)
}
}
