% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{code_template}
\alias{code_template}
\title{Code template for R code}
\usage{
code_template()
}
\value{
a character string with the code template.
}
\description{
Default code template with R code that is used when generating. For example,
variables in the template code are replaced with the actual values during
execution.
}
\details{
The following variables are placeholders and replaced with actual values:

\%\%FORMULA\%\% will be replaced with the lavaan formula.

\%\%OBSERVATIONS\%\% will be replaced with the number of observations of the
actual data.

\%\%TARGET\%\% will be replaced with the variable name of the covariance
matrix.

\%\%VARIABLES\%\% will be replaced with the column names (variable names) of
the covariance matrix.

\%\%VALUES\%\% will be replaced with the actual values of the covariance
matrix.
}
\examples{
\dontrun{
# Create random data.
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Modify code template.
template <- paste(code_template(), "\\n# A comment placed at end of code")

# Generate R code with the template.
code <- semproducible(data, formula="y ~ x", template = template)
}
}
