% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_template.r
\name{code_template}
\alias{code_template}
\title{Template for the generated R code}
\usage{
code_template()
}
\value{
a character string with the R code template.
}
\description{
Default R code template used by \code{\link{semproducible}} when generating
the code.
}
\details{
The template is a string with R code that contain placeholders in the form
of variables like \code{\%\%NAME\%\%}. The variables are replaced with
actual content of the SEM model during code generation.

Placeholder variables:

\itemize{
\item{\code{\%\%FORMULA\%\%}} = \code{\link[lavaan:sem]{lavaan}} formula.
\item{\code{\%\%OBSERVATIONS\%\%}} = number of observations of the actual data.
\item{\code{\%\%VARIABLES\%\%}} = column names (variable names) of
the covariance matrix.
\item{\code{\%\%VALUES\%\%}} = actual values of the covariance matrix.
\item{\code{\%\%LAVAAN_CALL\%\%}} = \code{\link[lavaan:sem]{lavaan}} function
call.
\item{\code{\%\%COVMAT_VARIABLE\%\%}} = variable name of the covariance matrix.
\item{\code{\%\%FIT_VARIABLE\%\%}} = variable name of the lavaan object that
is created from \code{\link[lavaan:sem]{lavaan}}.
}
}
\examples{
data <- data.frame(x = rnorm(100), y = rnorm(100))

code <- semproducible(data, formula="y ~ x", template = code_template())

# Modify code template
template <- paste(code_template(), "\n# Adds a comment at end")

code <- semproducible(data, formula="y ~ x", template = template)

# Look at the new code
cat(code)
}
