% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_code.R
\name{save_code}
\alias{save_code}
\title{Save generated code to a file}
\usage{
save_code(code, file, overwrite = FALSE)
}
\arguments{
\item{code}{character string with R code.}

\item{file}{character string of the filename to be saved.}

\item{overwrite}{whether the existing files should be overwritten.}
}
\description{
Helper function that saves generated R code to a file. Does not overwrite
existing files unless \code{overwrite} is explicitly set to \code{TRUE}.
}
\examples{
\dontrun{
data <- data.frame(x = rnorm(100), y = rnorm(100))

code <- semproducible(data, formula="y ~ x")

save_code(code, "create_data.R")

save_code(code, "create_data.R", overwrite = TRUE)
}
}
