% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_code.r
\name{save_code}
\alias{save_code}
\title{Save generated code to a file}
\usage{
save_code(code, file, overwrite = FALSE)
}
\arguments{
\item{code}{character string with R code.}

\item{file}{character string of the filename to be saved.}

\item{overwrite}{whether the existing files should be overwritten.}
}
\description{
Helper function that saves generated R code to a file. Does not overwrite
existing files unless \code{overwrite} is explicitly set to \code{TRUE}.
}
\examples{
\dontrun{
# Create random data.
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Generate R code.
code <- semproducible(data, formula="y ~ x")

# Save R code to file.
save_code(code, "create_data.r")

# Save again, but the file exists so overwrite=TRUE is necessary.
save_code(code, "create_data.r", overwrite = TRUE)
}
}
