% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{save_code}
\alias{save_code}
\title{Save code to a file}
\usage{
save_code(code, file, overwrite = FALSE)
}
\arguments{
\item{code}{character string with R code.}

\item{file}{character string of the filename to be saved.}

\item{overwrite}{whether the existing files should be overwritten.}
}
\description{
Save generated R code to a file.
}
\details{
Does not overwrite existing files unless
\code{overwrite} is set to \code{TRUE}.
}
\examples{
\dontrun{
# Create random data.
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Generate R code.
code <- semproducible(data, formula="y ~ x")

# Save R code to file.
save_code(code, "create_data.r")

# If the file already exists, you will get an error.
# You need to explicitly overwrite existing files.
save_code(code, "create_data.r", overwrite = TRUE)
}
}
