% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{save_code}
\alias{save_code}
\title{Save code to a file}
\usage{
save_code(code, filename, overwrite = FALSE)
}
\arguments{
\item{code}{R code to save (character).}

\item{filename}{filename of the file (character).}
}
\description{
Save code to a file
}
\examples{
\dontrun{
# Create random data.
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Generate R code.
code <- semproducible(data, formula="y ~ x")

# Save R code to file.
save_code(code, "create_data.r")

# If the file already exists, you will get an error.
# You need to explicitly overwrite existing files.
save_code(code, "create_data.r", overwrite = TRUE)
}
}
