% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{semproducible}
\alias{semproducible}
\title{Generate R code to reproduce a SEM model}
\usage{
semproducible(
  x,
  formula = NULL,
  target_variable = "cov_mat",
  digits = NULL,
  drop_non_numeric = FALSE,
  vars_per_line = 9,
  eval = FALSE,
  use = "complete.obs",
  template = NULL
)
}
\arguments{
\item{x}{a \code{\link[base:data.frame]{data.frame}} with numeric variables,
a \code{\link[base:matrix]{matrix}} with covariances, or a
\code{\link[lavaan:sem]{lavaan}} object.}

\item{formula}{character string with a custom lavaan formula syntax
(e.g. \code{y ~ x}) that should be included in the code. If formula is
\code{NULL}, and \code{x} is a lavaan object, then a formula is created from
the lavaan object.}

\item{target_variable}{character string with arbitrary target variable name
for the generated covariance matrix. Defaults to \code{cov_mat}.}

\item{digits}{number of decimal digits of the covariance matrix. The
default (\code{NULL}) will show all available digits as specified by R
options. The higher the number of decimal digits, the more accurate the
reproducible model will be.}

\item{drop_non_numeric}{whether non-numeric columns should be dropped
from the data frame. This is useful if you have characters of factors as
columns which should not be included in the covariance matrix. Defaults to
\code{FALSE}.}

\item{vars_per_line}{number of variables/values per line. Use a low
value to decrease the width of the generated code.}

\item{eval}{whether the generated code and \code{\link[lavaan:sem]{lavaan}}
model will be executed during generation. If \code{TRUE}, a message will
tell you whether the code executed without errors or not.}

\item{use}{character string for computing the covariances in the presence
of missing values. This must be \code{everything}, \code{all.obs},
\code{complete.obs}, \code{na.or.complete}, or \code{pairwise.complete.obs}.
This value is passed on to the \code{use} parameter of the
\code{\link[stats:cor]{cov}} function.}

\item{template}{a character string with a custom code template that
is used when generating the R code. A \code{NULL} value (default) will use
the default template from \code{\link{code_template}}.}
}
\value{
a character string with the generated R code.
}
\description{
Generate R code from your data frame, an existing covariance matrix, or
a \code{lavaan} object. The R code reproduces your structural equation model
(SEM) to support open science with minimal effort, and makes it easy to
supply your code to a journal article, or to share your model with others
without sharing your full dataset that may contain sensitive information.
}
\details{

}
\examples{
\dontrun{
library(semproducible)
library(lavaan)

# Create random data.
set.seed(5543)
data <- data.frame(x = rnorm(100),
                   y = rnorm(100),
                   z = rnorm(100),
                   w = rnorm(100),
                   q = rnorm(100))

# Generate R code for the data.
code <- semproducible(data, formula="y ~ x")

# Use 5 decimal digits.
code <- semproducible(data, formula="y ~ x", digits=5)

# Restrict number of values per line to 4.
code <- semproducible(data, formula="y ~ x", digits=5, vars_per_line=4)

# View code.
cat(code)

# Save code to a file.
save_code(code, "create_data.r")

# Reproduce lavaan object (from http://lavaan.ugent.be/tutorial/cfa.html)
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '
fit <- cfa(HS.model, data=HolzingerSwineford1939)
code <- semproducible(fit)
}
}
