% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{semproducible}
\alias{semproducible}
\title{Code generator for reproducible covariance matrix}
\usage{
semproducible(x, digits = NULL, use = "complete.obs",
  formula = "YOUR lavaan MODEL HERE", target_variable = "cov_mat",
  drop_non_numeric = FALSE, vars_per_line = 9, eval = FALSE,
  print = FALSE, template = NULL)
}
\arguments{
\item{x}{a data frame with numeric variables, or a matrix with correlations
or covariances.}

\item{digits}{number of decimal digits. The default (NULL) will show all
available digits as specified by R options. The higher the number of
decimal digits, the more accurate the reproducible model will be.}

\item{use}{optional character string giving a method for computing
covariances in the presence of missing values. This must be one of the
strings "everything", "all.obs", "complete.obs", "na.or.complete", or
"pairwise.complete.obs". This value is passed on to the cov() method.}

\item{formula}{optional argument that specifies the lavaan formula syntax
that should be included in the code.}

\item{target_variable}{target variable name for the generated covariance
matrix. Defaults to "cov_mat".}

\item{drop_non_numeric}{drop columns from the data frame that are not
numeric. This is useful if you have characters of factors as columns which
should not be included in the covariance matrix. Defaults to FALSE.}

\item{vars_per_line}{number of variables (or values) per line. Many
variables per line will increase the width of the code.}

\item{eval}{whether or not the generated code and lavaan model will be
executed. If eval is set to TRUE, a message will tell you whether the
code executed withour errors or not.}

\item{print}{whether or not to print the code to the screen or return the
code as a character string (default). Printing to screen is useful during
development.}

\item{template}{a character string with a custom code template that
is used when generating the R code. See the code_template() function for
instructions on how to write your own template. A NULL value (default)
will use the default template from code_template().}
}
\value{
a character string with the generated R code to reproduce the
covariance matrix and the necessary lavaan code to run it.
}
\description{
Generate R code from your data frame or an existing covariance matrix.
This is useful when you need to create a reproducible covariance matrix
that can be used by a structural equation model (SEM) in lavaan.
}
\details{
You supply a data frame with numeric variables (or a covariance matrix that
you have already prepared). The function then generates R code that
can reproduce the SEM model as a covariance matrix.
}
\examples{
\dontrun{
# Create random data.
set.seed(5543)
data <- data.frame(x = rnorm(100),
                   y = rnorm(100),
                   z = rnorm(100),
                   w = rnorm(100),
                   q = rnorm(100))

# Generate R code for the data.
cat(semproducible(data, formula="y ~ x"))

# Use 5 decimal digits.
cat(semproducible(data, formula="y ~ x", digits=5))

# Restrict number of values per line to 4.
cat(semproducible(data, formula="y ~ x", digits=5, vars_per_line=4))

# Save code to a file.
code <- semproducible(data, formula="y ~ x")
save_code(code, "create_data.r")
}
}
