% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{semproducible}
\alias{semproducible}
\title{Generate R code to reproduce a latent variable model (lavaan)}
\usage{
semproducible(
  x,
  formula = NULL,
  covmat_variable = "cov_mat",
  fit_variable = "fit",
  digits = NULL,
  drop_non_numeric = FALSE,
  comments = TRUE,
  vars_per_line = 9,
  eval = FALSE,
  use = "complete.obs",
  template = code_template()
)
}
\arguments{
\item{x}{a \code{\link[base:data.frame]{data.frame}} with numeric variables,
a \code{\link[base:matrix]{matrix}} with covariances, or a fitted
\code{\link[lavaan:sem]{lavaan}} object that you want to reproduce.}

\item{formula}{character string with a custom lavaan formula syntax
(e.g. \code{y ~ x}) that should be included in the code. If formula is
\code{NULL}, and \code{x} is a lavaan object, then a formula is generated
automatically from the lavaan object.}

\item{covmat_variable}{character string with arbitrary variable name
for the generated covariance matrix. Defaults to \code{"cov_mat"}.}

\item{fit_variable}{character string with arbitrary target variable name
for the lavaan object. Defaults to \code{"fit"}.}

\item{digits}{number of decimal digits of the covariance matrix. The
default (\code{NULL}) will show all available digits as specified by your R
options. The higher the number of decimal digits, the more accurate the
reproducible model will be.}

\item{drop_non_numeric}{whether non-numeric columns should be dropped
from the data frame. This is useful if you have characters or factors as
columns, which should0 be excluded from the covariance matrix. Defaults to
\code{FALSE}.}

\item{comments}{whether the generated code should include comments that
describe the code. Defaults to TRUE.}

\item{vars_per_line}{number of variables/values per line. Use a low
value to decrease the width of the generated code.}

\item{eval}{whether the generated code and \code{\link[lavaan:sem]{lavaan}}
model will be executed during generation. If \code{TRUE}, a message will
tell you whether the code executed with errors or not.}

\item{use}{character string for computing the covariances in the presence
of missing values. This value is simply passed on to the \code{use}
parameter of the \code{\link[stats:cor]{cov}} function. The value must be
\code{everything}, \code{all.obs}, \code{complete.obs},
\code{na.or.complete}, or \code{pairwise.complete.obs}.}

\item{template}{a character string with a custom code template that
is used when generating the R code. The default value is
\code{\link{code_template}} which contains code for lavaan.}
}
\value{
a character string with the generated R code.
}
\description{
Generate R code from your data frame, an existing covariance matrix, or
a \code{lavaan} object.
}
\details{
The R code reproduces your structural equation model
(SEM) to support open science with minimal effort, and makes it easy to
supply your code to a journal article, or to share your model with others
without sharing your full dataset that may contain sensitive information.

Semproducible is useful when you need to create a reproducible covariance
matrix that can be used by a structural equation model (SEM) in
\code{\link[lavaan:sem]{lavaan}}.

You supply a data frame with numeric variables (or a covariance matrix that
you have already prepared). \code{semproducible} then generates R code that
can reproduce the SEM model as a covariance matrix.

You can also directly supply a lavaan SEM model, and semproducible will use
the fitted (observed) covariance matrix and produce R code that reproduce the
model, using the estimator of your lavaan model. Note, however, that this
will not give you all the possible models that you could have ran with all
the data, but only the variables that are passed to lavaan.

Save the code to a file using the
\code{\link[semproducible:save_code]{save_code}} function.
}
\examples{
\dontrun{
library(semproducible)
library(lavaan)

# Create random data
set.seed(5543)
data <- data.frame(x = rnorm(100),
                   y = rnorm(100),
                   z = rnorm(100),
                   w = rnorm(100),
                   q = rnorm(100))

code <- semproducible(data, formula="y ~ x")

code <- semproducible(data, formula="y ~ x", digits=5)

code <- semproducible(data, formula="y ~ x", digits=5, vars_per_line=4)

# View code
cat(code)

save_code(code, "create_data.r")

# Example: http://lavaan.ugent.be/tutorial/cfa.html
HS.model <- ' visual  =~ x1 + x2 + x3
              textual =~ x4 + x5 + x6
              speed   =~ x7 + x8 + x9 '
fit <- cfa(HS.model, data=HolzingerSwineford1939)
code <- semproducible(fit)
}
}
