% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semproducible.r
\name{semproducible}
\alias{semproducible}
\title{Code generator for reproducible covariance matrix}
\usage{
semproducible(x, digits = NULL, early_line_break = FALSE,
  method = "cov", formula = "YOUR lavaan MODEL HERE",
  target_variable = "data", drop_non_numeric = FALSE)
}
\arguments{
\item{x}{a data frame with numeric variables, or a matrix with correlations
or covariances.}

\item{digits}{number of decimal digits. The default (NULL) will show all
available digits as specified by R options. The higher the number of
decimal digits, the more accurate the reproducible model will be.}

\item{early_line_break}{use early line break. This will show a single value
per line instead of several values per line. This is only of importance
if you plan to print your code or include it in a manuscript or article
with a fixed width. Defaults to FALSE.}

\item{method}{method for creating the matrix, either "cor" for correlations
or "cov" for covariances (default). This argument is ignored if "x"
is a matrix.}

\item{formula}{optional argument that specifies the lavaan formula syntax
that should be included in the code.}

\item{target_variable}{target variable name for the generated code. Defaults
to "data", but can be anything you want.}

\item{drop_non_numeric}{drop columns from the data frame that are not
numeric. This is useful if you have characters of factors as columns which
should not be included in the covariance matrix. Defaults to FALSE.}
}
\value{
Character with the generated R code to reproduce the covariance
matrix and the necessary lavaan code to run it.
}
\description{
Generate R code from your data frame or an existing covariance matrix.
This is useful when you need to create a reproducible covariance matrix
that can be used by a structural equation model (SEM) in lavaan.
}
\details{
You supply a data frame with numeric variables (or a covariance matrix that
you have already prepared). The function then generates R code that
can reproduce the SEM model as a correlation matrix or a covariance matrix.
}
\examples{
\dontrun{
# Create random data.
set.seed(5543)
data <- data.frame(x = rnorm(100),
                   y = rnorm(100),
                   z = rnorm(100),
                   w = rnorm(100),
                   q = rnorm(100))

# Generate R code for the data.
cat(semproducible(data, formula="y ~ x"))

# Generate R code for the data, with 5 decimal digits.
cat(semproducible(data, formula="y ~ x", digits=5))

# Generate the same code, but keep each value on one line for easier reading.
cat(semproducible(data, formula="y ~ x", digits=5, early_line_break=TRUE))

# Save R code to a file.
code <- semproducible(data, formula="y ~ x")
fileConn <- file("create_data.r")
writeLines(code, fileConn)
close(fileConn)
}
}
